+++ example/a.ts
@@ -1,21 +1,24 @@
-import { Elysia, error, t } from '../src'
+import { Elysia, t } from '../src'
 import { req } from '../test/utils'
 
 const app = new Elysia().get(
 	'/',
 	() => {
 		return {
-			name: 'a',
-			a: 'b'
+			hello: 'world'
 		}
 	},
 	{
 		response: t.Object({
-			name: t.String()
+			hello: t.String({
+				sanitize: true
+			})
 		})
 	}
 )
 
-app.handle(req('/'))
-	.then((x) => x.json())
-	.then(console.log)
+const response = await app.handle(req('/')).then((x) => x.json())
+
+console.log(response)
+
+console.log(app.routes[0].compile().toString())