+++ src/types.ts
@@ -35,6 +35,7 @@ import type { ComposerGeneralHandlerOptions } from './compose'
 import type { ElysiaAdapter } from './adapter'
 import type { AnyWSLocalHook, WSLocalHook } from './ws/types'
 import type { WebSocketHandler } from './ws/bun'
+import { ElysiaTypeCheck } from './schema'
 
 type PartialServe = Partial<Serve>
 
@@ -170,6 +171,22 @@ export type ElysiaConfig<Prefix extends string | undefined> = {
 	 * @since 1.1.11
 	 */
 	nativeStaticResponse?: boolean
+	/**
+	 * Use runtime/framework provided router if possible
+	 *
+	 * @default true
+	 * @since 1.3.0
+	 */
+	systemRouter?: boolean
+	/**
+	 * When response schema is provided
+	 * Use Elysia custom JSON encoder to optimize encode process
+	 *
+	 *
+	 * @default true
+	 * @since 1.3.0
+	 */
+	jsonAccelerator?: boolean
 }
 
 export type ValidatorLayer = {
@@ -480,6 +497,8 @@ export type HookContainer<T extends Function = Function> = {
 	scope?: LifeCycleType
 	subType?: 'derive' | 'resolve' | 'mapDerive' | 'mapResolve' | (string & {})
 	fn: T
+	isAsync?: boolean
+	hasReturn?: boolean
 }
 
 export interface LifeCycleStore {
@@ -1250,24 +1269,25 @@ export interface InternalRoute {
 	method: HTTPMethod
 	path: string
 	composed: ComposedHandler | Response | null
+	compile(): ComposedHandler
 	handler: Handler
 	hooks: AnyLocalHook
 	websocket?: AnyWSLocalHook
 }
 
 export type SchemaValidator = {
-	createBody?(): TypeCheck<any>
-	createHeaders?(): TypeCheck<any>
-	createQuery?(): TypeCheck<any>
-	createParams?(): TypeCheck<any>
-	createCookie?(): TypeCheck<any>
-	createResponse?(): Record<number, TypeCheck<any>>
-	body?: TypeCheck<any>
-	headers?: TypeCheck<any>
-	query?: TypeCheck<any>
-	params?: TypeCheck<any>
-	cookie?: TypeCheck<any>
-	response?: Record<number, TypeCheck<any>>
+	createBody?(): ElysiaTypeCheck<any>
+	createHeaders?(): ElysiaTypeCheck<any>
+	createQuery?(): ElysiaTypeCheck<any>
+	createParams?(): ElysiaTypeCheck<any>
+	createCookie?(): ElysiaTypeCheck<any>
+	createResponse?(): Record<number, ElysiaTypeCheck<any>>
+	body?: ElysiaTypeCheck<any>
+	headers?: ElysiaTypeCheck<any>
+	query?: ElysiaTypeCheck<any>
+	params?: ElysiaTypeCheck<any>
+	cookie?: ElysiaTypeCheck<any>
+	response?: Record<number, ElysiaTypeCheck<any>>
 }
 
 export type AddPrefix<Prefix extends string, T> = {
@@ -1612,9 +1632,13 @@ export type MergeElysiaInstances<
 					Definitions['error'] &
 						Current['_types']['Definitions']['error']
 				>
-				typebox: MergeTypeModule<
-					Definitions['typebox'],
-					Current['_types']['Definitions']['typebox']
+				typebox: TModule<
+					Prettify<
+						UnwrapTypeModule<Definitions['typebox']> &
+							UnwrapTypeModule<
+								Current['_types']['Definitions']['typebox']
+							>
+					>
 				>
 			},
 			Metadata & Current['_types']['Metadata'],