+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 1.3.0
+Feature:
+- add `jsonAccelerator`
+- add `systemRouter` config
+- add sucrose cache
+
+Improvement:
+- reduce redundant type check when using Encode
+- optimize isAsync
+
 # 1.2.24 - 2 Mar 2025
 Bug fix:
 - 200 object response is not inferring type in type-level